# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Awards Management API",
        "description": "API for managing awards, nominees, voting, and events",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Awards Management API Server"
        }
    ],
    "paths": {
...
        "/api/v1/votes/{nominee_id}": {
            "post": {
                "tags": [
                    "Voting"
                ],
                "summary": "Vote for a nominee",
                "description": "Cast a vote for a specific nominee",
                "operationId": "voteForNominee",
                "parameters": [
                    {
                        "name": "nominee_id",
                        "in": "path",
                        "description": "Nominee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ip_address": {
                                        "description": "Voter IP address",
                                        "type": "string",
                                        "example": "192.168.1.1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vote recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vote recorded successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "nominee_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "new_vote_count": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "new_percentage": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Voting is currently closed for this category."
                                        },
                                        "errors": {
                                            "properties": {
                                                "voting": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "You have already voted for this nominee."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nominee not found"
                    }
                }
            }
        }
    },
...
}